# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "keeper project",
        "title": "Keeper Server",
        "contact": {
            "name": "Nexbug",
            "url": "https://github.com/nextlag",
            "email": "nextbug@ya.ru"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/user/binary": {
            "get": {
                "description": "Retrieve all binaries uploaded by the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "binaries"
                ],
                "summary": "Get all binaries for the current user",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Binary"
                            }
                        }
                    },
                    "204": {
                        "description": "No content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            },
            "post": {
                "description": "Upload a new binary for the current user",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "binaries"
                ],
                "summary": "Add a new binary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Binary name",
                        "name": "name",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Binary file",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/entity.Binary"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
        "/user/binary/{id}": {
            "get": {
                "description": "Download a specific binary identified by its UUID",
                "tags": [
                    "binaries"
                ],
                "summary": "Download a binary by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Binary UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a specific binary identified by its UUID",
                "tags": [
                    "binaries"
                ],
                "summary": "Delete a binary by UUID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Binary UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "delete accepted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
        "/user/binary/{id}/meta": {
            "post": {
                "description": "Add metadata to a specific binary identified by its UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "binaries"
                ],
                "summary": "Add metadata to a binary",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Binary UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Metadata for the binary",
                        "name": "metadata",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Meta"
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.Meta"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "entity.Binary": {
            "type": "object",
            "properties": {
                "file_name": {
                    "description": "Filesystem name.",
                    "type": "string"
                },
                "meta": {
                    "description": "Associated metadata.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.Meta"
                    }
                },
                "name": {
                    "description": "File name.",
                    "type": "string"
                }
            }
        },
...
        "entity.Meta": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "Name or type of the metadata.",
                    "type": "string"
                },
                "value": {
                    "description": "Value of the metadata.",
                    "type": "string"
                }
            }
        },
...
        "v1.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                },
                "status": {
                    "type": "string",
                    "example": "message"
                }
            }
        }
    }
}